+++ akka-actor/src/main/scala/akka/routing/Routing.scala
@@ -6,6 +6,7 @@ package akka.routing
 
 import akka.AkkaException
 import akka.actor._
+import akka.util.Timeout
 import akka.config.ConfigurationException
 import akka.dispatch.{ Future, MessageDispatcher }
 import akka.util.{ ReflectiveAccess, Duration }
